Data now returns raw bytes transfered. It is no longer fixed at KBps and leaves the caller to determine how it should be represented.

+++ windows/monitor-dll/monitor/monitor.h
@@ -4,26 +4,24 @@
 #include <string>
 #include <memory>
 
+#include <stdint.h>
+
 #if defined _MSC_VER
 // Disable: 'identifier' : class 'type' needs to have dll-interface to be used by clients of class 'type2' ('std::unique_ptr<_Ty>', 'Monitor')
 #pragma warning(disable:4251)
 #endif
 
 struct __declspec(dllexport) Data {
-    double rx;
-    double tx;
+    uint64_t rx;
+    uint64_t tx;
 };
 
 class __declspec(dllexport) Monitor {
 public:
     Monitor(size_t = 0); // Default interface_id value of 0 [All Interfaces (*)]
-
     ~Monitor();
-
     Data poll();
-
     std::vector<std::wstring> interfaces();
-
     void switch_interface(size_t);
 
 private:
